% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution_functions.R
\name{pgev}
\alias{pgev}
\alias{dgev}
\alias{qgev}
\alias{rgev}
\title{Generalized Extreme Value distribution}
\usage{
pgev(q, loc = 0, scale = 1, shape = 0)

dgev(x, loc = 0, scale = 1, shape = 0)

qgev(p, loc = 0, scale = 1, shape = 0)

rgev(n, loc = 0, scale = 1, shape = 0)
}
\arguments{
\item{loc, scale, shape}{location, scale, and shape parameter of the generalized extreme value distribution. All must be of length one.}

\item{x, q, p}{numeric vector of values, quantiles, or probabilites.}

\item{n}{numeric, number of random variates.}
}
\description{
Cumulative distribution function, density function, quantile function and
generation of random variates of the generalized extreme value distribution.
}
\seealso{
\code{\link{pgum}}, \code{\link{pgpd}}, \code{\link{pln3}}
}
