% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution_functions.R
\name{pgpd}
\alias{pgpd}
\alias{dgpd}
\alias{qgpd}
\alias{rgpd}
\title{Generalized Pareto distribution}
\usage{
pgpd(q, loc = 0, scale = 1, shape = 0)

dgpd(x, loc = 0, scale = 1, shape = 0)

qgpd(p, loc = 0, scale = 1, shape = 0)

rgpd(n, loc = 0, scale = 1, shape = 0)
}
\arguments{
\item{loc, scale, shape}{location, scale, and shape parameter of the generalized Pareto distribution. All must be of length one.}

\item{x, q, p}{numeric vector of values, quantiles, or probabilites.}

\item{n}{numeric, number of random variates.}
}
\description{
Cumulative distribution function, density function, quantile function and
generation of random variates of the generalized Pareto distribution.
}
\seealso{
\code{\link{pgev}}, \code{\link{pgum}}, \code{\link{pln3}}
}
