% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution_functions.R
\name{pgum}
\alias{pgum}
\alias{dgum}
\alias{qgum}
\alias{rgum}
\title{Gumbel distribution}
\usage{
pgum(q, loc = 0, scale = 1)

dgum(x, loc = 0, scale = 1)

qgum(p, loc = 0, scale = 1)

rgum(n, loc = 0, scale = 1)
}
\arguments{
\item{loc, scale}{location and scale parameter of the Gumbel distribution. All must be of length one.}

\item{x, q, p}{numeric vector of values, quantiles, or probabilites.}

\item{n}{numeric, number of random variates.}
}
\description{
Cumulative distribution function, density function, quantile function and
generation of random variates of the Gumbel distribution.
}
\seealso{
\code{\link{pgev}}, \code{\link{pgpd}}, \code{\link{pln3}}
}
