% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution_functions.R
\name{pln3}
\alias{pln3}
\alias{dln3}
\alias{qln3}
\alias{rln3}
\title{Three-parameter lognormal distribution}
\usage{
pln3(q, loc = 0, scale = 1, shape = 1)

dln3(x, loc = 0, scale = 1, shape = 1)

qln3(p, loc = 0, scale = 1, shape = 1)

rln3(n, loc = 0, scale = 1, shape = 1)
}
\arguments{
\item{loc, scale, shape}{location, scale, and shape parameter of the three-parameter lognormal distribution. All must be of length one.}

\item{x, q, p}{numeric vector of values, quantiles, or probabilites.}

\item{n}{numeric, number of random variates.}
}
\description{
Cumulative distribution function, density function, quantile function and
generation of random variates of the three-parameter lognormal distribution.
}
\seealso{
\code{\link{pgev}}, \code{\link{pgum}}, \code{\link{pgpd}}
}
