% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{returnParameters}
\alias{returnParameters}
\title{returnParameters}
\usage{
returnParameters(out, distribution, ...)
}
\arguments{
\item{out}{-}

\item{distribution}{-}

\item{...}{-}
}
\value{
An object of class parameters.
}
\description{
Sets attributions to parameters objects and returns them. This function is for internal use.
}
