% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PWMs.R
\name{returnPWMs}
\alias{returnPWMs}
\title{returnPWMs}
\usage{
returnPWMs(out, order, ...)
}
\arguments{
\item{out}{-}

\item{order}{-}

\item{...}{-}
}
\value{
An object of class PWMs.
}
\description{
Sets attributes to PWMs objects and returns them. This function is for internal use.
}
