% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantiles.R
\name{returnQuantiles}
\alias{returnQuantiles}
\title{returnQuantiles}
\usage{
returnQuantiles(out, distribution, p, ...)
}
\arguments{
\item{out}{-}

\item{distribution}{-}

\item{p}{-}

\item{...}{-}
}
\value{
An object of class quantiles.
}
\description{
Sets attributions to quantiles objects and returns them. This function is for internal use.
}
