% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TLMoments.R
\name{returnTLMoments}
\alias{returnTLMoments}
\title{returnTLMoments}
\usage{
returnTLMoments(out, leftrim, rightrim, order, ...)
}
\arguments{
\item{out}{-}

\item{leftrim}{-}

\item{rightrim}{-}

\item{order}{-}

\item{...}{-}
}
\value{
An object of class TLMoments.
}
\description{
Sets attributions to TLMoments objects and returns them. This function is for internal use.
}
